+++ src/test/kotlin/Day7AmplificationCircuitTest.kt
@@ -14,7 +14,7 @@ internal class Day7AmplificationCircuitTest {
         "3,31,3,32,1002,32,10,32,1001,31,-2,31,1007,31,0,33,1002,33,7,33,1,33,31,31,1,32,31,31,4,31,99,0,0,0;65210"],
         delimiter = ';')
     fun maxThrusterSignalExamples(program: String, expected: Int) {
-        val programArray = program.split(",").stream().mapToInt { it.toInt() }.toArray()
+        val programArray = program.split(",").stream().mapToLong() { it.toLong() }.toArray()
         Assertions.assertEquals(expected, Day7AmplificationCircuit().maxThrusterSignal(programArray))
     }
 
@@ -22,7 +22,7 @@ internal class Day7AmplificationCircuitTest {
     fun maxThrusterSignal() {
         val programArray = Files.lines(Paths.get("./src/test/resources/day7.txt"))
             .flatMap { it.split(",").stream() }
-            .mapToInt { it.toInt() }
+            .mapToLong { it.toLong() }
             .toArray()
         Assertions.assertEquals(65464, Day7AmplificationCircuit().maxThrusterSignal(programArray))
     }
